requirements.txt
openai>=1.14.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0
typer[all]>=0.9.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.0
rich>=13.7.0
